backend/requirements.txt
Django==4.2.7
psycopg2-binary==2.9.9
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
requests==2.31.0
django-cors-headers==4.3.1
gunicorn==21.2.0
